% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcess.R
\name{addCapitalServices}
\alias{addCapitalServices}
\title{Calculate the capital services, both with user costs in CP and with mid
period productive stock in reference year user cost prices. A by-product is
the return to capital.}
\usage{
addCapitalServices(capStockData, referenceYear, discountRate)
}
\arguments{
\item{referenceYear}{a numeric value indicating the year against which all prices are set}

\item{discountRate}{a numeric value between [0,1]}

\item{capStockdata}{data.frame with productive stock in CVM (ProductiveStock),
the price index (PriceIndex), tax adjustment factor (TaxAdjustmentFactor),
productive stock utilisation rate (UTIL),
consumption of fixed capital in cp (ConsumptionOFixedCapitalCP) and
net stock in CVM (NetStock)}
}
\value{
the given data.frame with the addition of return to capital, capital
services in CP and capital service in CVM
}
\description{
Calculate the capital services, both with user costs in CP and with mid
period productive stock in reference year user cost prices. A by-product is
the return to capital.
}
