% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcess.R
\name{addHoldingGainsLosses}
\alias{addHoldingGainsLosses}
\title{Calculate nominal/real/neutral holding gains/losses}
\usage{
addHoldingGainsLosses(capStockData)
}
\arguments{
\item{capStockData}{data.frame with net stock in CP (NetStockCP), net fixed
capital formation in CP (NetFixedCapitalFormationCP), total other changes in
volume in CP (TotalOtherChangesInVolumeCP), the price index (PriceIndex) and
the \% growth in consumer prices index/100 (CPI)}
}
\value{
the given data frame with the addition of nominal/real/neutral
holding gains/losses
}
\description{
Calculate nominal/real/neutral holding gains/losses
}
