% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPriceIndex.R
\name{calcPriceIndexUsingAvgs}
\alias{calcPriceIndexUsingAvgs}
\title{Calculate Price Index using Averages}
\usage{
calcPriceIndexUsingAvgs(dat)
}
\arguments{
\item{dat}{one or more numeric values in current prices - Updated 15-11-18}
}
\value{
a numeric vector
}
\description{
Calculate the price index as the difference between values in current prices
and values in chained volume measures
Then create an average of these based on at least 2 quarters from the current year and 2 quarters from thwe previous/next year
Example
2010Q1 = (2009Q3 + 2009Q4 + 2010Q1 + 2010Q2) / 4
2010Q2 = (2009Q3 + 2009Q4 + 2010Q1 + 2010Q2) / 4
2010Q3 = (2010Q3 + 2010Q4 + 2011Q1 + 2011Q2) / 4
2010Q4 = (2010Q3 + 2010Q4 + 2011Q1 + 2011Q2) / 4
}
\details{
CP & CVM values are BACKCAST 5 Periods and FORECAST 5 Periods then CP divided by CVM to create the orignal index.
This then allows you to take the average for the earliest and latest Periods
}
\examples{
dat(2:12, 2:12)
}
