% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCPtoCVM.R
\name{convertCPToCVM}
\alias{convertCPToCVM}
\title{Convert Prices from CVM to CP}
\usage{
convertCPToCVM(cpData, priceIndex)
}
\arguments{
\item{cpData}{one or more values in current prices}

\item{priceIndex}{one or more values for the price index}
}
\value{
a numeric vector
}
\description{
Convert data that is current prices to data that is chained volume measures
by dividing by a price index
}
\examples{
convertCPToCVM(20, 0.69)
convertCPToCVM(10:20, rep(0.5, 10))
}
