% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCVMToCP.R
\name{convertCVMToCP}
\alias{convertCVMToCP}
\title{Convert Prices from CVM to CP}
\usage{
convertCVMToCP(cvmData, priceIndex)
}
\arguments{
\item{cvmData}{one or more values in current prices}

\item{priceIndex}{one or more values for the price index}
}
\value{
a numeric vector
}
\description{
Convert data that is in chained volume measures prices to data that is in
current prices by multiplying with a price index
}
\examples{
convertCVMToCP(20, 0.69)
convertCVMToCP(10:20, rep(0.5, 10))
}
