% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runAll}
\alias{runAll}
\title{Run all.}
\usage{
runAll(.data, parallelise = TRUE)
}
\arguments{
\item{.data}{A data.frame with one row per Sector/Asset/Industry with a "data"
list-column containing GFCF, a "config" list-column, and a refYear column. See details below.}

\item{parallelise}{Set to FALSE to turn off parallel processing.}
}
\value{
The input data.frame with an addional "result" list-column containing the PIM results.
}
\description{
Run all series through the PIM.
}
\section{Data Specification}{

The .data input data.frame should consist of one row per Sector/Industry/Asset
with a nested "data" column containing the data for that series, and a nested
"config" column containing the output from pimir::pimConfig for that series.
For example, have the following structure:

\tabular{rrrrrr}{
Sector \tab Industry \tab Asset \tab data \tab config \tab refYear\cr
S.11001 \tab 02 \tab OTHER.BUILDINGS \tab <data.frame> \tab <list> \tab Y2014 \cr
S.11001 \tab 02 \tab OTHER.STRUCTURES \tab <data.frame> \tab <list> \tab Y2014 \cr
... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}

The nested "data" column should have the following structure:

\tabular{lrrrrrrrrrrrrrrrr}{
Vintage \tab gfcfCP \tab PriceIndex \tab Average \tab CoV \tab Max \tab Min \tab K1CP \tab K3CP \tab K4CP \tab K5CP \tab K61CP \tab K62CP \tab CPI \tab TaxAdjustmentFactor \tab UTIL  \cr
Y1948Q1 \tab 10 \tab 0.050 \tab 80 \tab 0.15 \tab 128 \tab 4 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0.04 \tab 1.04 \tab 1 \cr
Y1948Q2 \tab 12 \tab 0.057 \tab 80 \tab 0.15 \tab 128 \tab 4 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0 \tab 0.04 \tab 1.04 \tab 1 \cr
... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}

The data.frame should also contain a "refYear" column used in the calculation of
Capital Services.
}

